clean distinction between guest and user home pages
