+++ boardgame-ui/src/pages/GuestHomePage.vue
@@ -1,0 +1,13 @@
+<template>
+    Hello
+</template>
+
+<script>
+export default {
+    name: "UserHomePage"
+}
+</script>
+
+<style scoped>
+
+</style>
